@@ -1,17 +1,19 @@
-const {h, Component} = require('preact')
-const classNames = require('classnames')
+import {h, Component} from 'preact'
+import classNames from 'classnames'
 
-class InfoOverlay extends Component {
-    shouldComponentUpdate({text, show}) {
-        return text !== this.props.text || show !== this.props.show
-    }
+export default class InfoOverlay extends Component {
+  shouldComponentUpdate({text, show}) {
+    return text !== this.props.text || show !== this.props.show
+  }
 
-    render({text, show}) {
-        return h('section', {
-            id: 'info-overlay',
-            class: classNames({show})
-        }, text)
-    }
+  render({text, show}) {
+    return h(
+      'section',
+      {
+        id: 'info-overlay',
+        class: classNames({show})
+      },
+      text
+    )
+  }
 }
-
-module.exports = InfoOverlay
