@@ -1,127 +1,193 @@
-const {remote} = require('electron')
-const {h, Component} = require('preact')
-const classNames = require('classnames')
+import * as remote from '@electron/remote'
+import {h, Component} from 'preact'
+import classNames from 'classnames'
 
-const Drawer = require('./Drawer')
+import i18n from '../../i18n.js'
+import sabaki from '../../modules/sabaki.js'
+import {noop, getScore} from '../../modules/helper.js'
 
-const t = require('../../i18n').context('ScoreDrawer')
-const helper = require('../../modules/helper')
+import Drawer from './Drawer.js'
+
+const t = i18n.context('ScoreDrawer')
 const setting = remote.require('./setting')
 
 class ScoreRow extends Component {
-    render({method, score, komi, handicap, sign}) {
-        let index = sign > 0 ? 0 : 1
+  render({method, score, komi, handicap, sign}) {
+    let index = sign > 0 ? 0 : 1
 
-        let total = !score ? 0 : method === 'area' ? score.area[index]
-            : score.territory[index] + score.captures[index]
+    let total = !score
+      ? 0
+      : method === 'area'
+      ? score.area[index]
+      : score.territory[index] + score.captures[index]
 
-        if (sign < 0) total += komi
-        if (method === 'area' && sign < 0) total += handicap
+    if (sign < 0) total += komi
+    if (method === 'area' && sign < 0) total += handicap
 
-        return h('tr', {},
-            h('th', {},
-                h('img', {
-                    src: `./node_modules/@sabaki/shudan/css/stone_${sign}.png`,
-                    alt: sign > 0 ? t('Black') : t('White'),
-                    width: 24,
-                    height: 24
-                })
-            ),
-            h('td', {class: classNames({disabled: method === 'territory'})}, score ? score.area[index] : '-'),
-            h('td', {class: classNames({disabled: method === 'area'})}, score ? score.territory[index] : '-'),
-            h('td', {class: classNames({disabled: method === 'area'})}, score ? score.captures[index] : '-'),
-            h('td', {}, sign < 0 ? komi : '-'),
-            h('td', {class: classNames({disabled: method === 'territory'})}, sign < 0 ? handicap : '-'),
-            h('td', {}, total)
-        )
-    }
+    return h(
+      'tr',
+      {},
+      h(
+        'th',
+        {},
+        h('img', {
+          src: `./node_modules/@sabaki/shudan/css/stone_${sign}.svg`,
+          alt: sign > 0 ? t('Black') : t('White'),
+          width: 24,
+          height: 24
+        })
+      ),
+      h(
+        'td',
+        {class: classNames({disabled: method === 'territory'})},
+        score ? score.area[index] : '-'
+      ),
+      h(
+        'td',
+        {class: classNames({disabled: method === 'area'})},
+        score ? score.territory[index] : '-'
+      ),
+      h(
+        'td',
+        {class: classNames({disabled: method === 'area'})},
+        score ? score.captures[index] : '-'
+      ),
+      h('td', {}, sign < 0 ? komi : '-'),
+      h(
+        'td',
+        {class: classNames({disabled: method === 'territory'})},
+        sign < 0 ? handicap : '-'
+      ),
+      h('td', {}, total)
+    )
+  }
 }
 
-class ScoreDrawer extends Component {
-    constructor() {
-        super()
+export default class ScoreDrawer extends Component {
+  constructor() {
+    super()
 
-        this.handleTerritoryButtonClick = () => setting.set('scoring.method', 'territory')
-        this.handleAreaButtonClick = () => setting.set('scoring.method', 'area')
-        this.handleCloseButtonClick = () => sabaki.closeDrawer()
+    this.handleTerritoryButtonClick = () =>
+      setting.set('scoring.method', 'territory')
+    this.handleAreaButtonClick = () => setting.set('scoring.method', 'area')
+    this.handleCloseButtonClick = () => sabaki.closeDrawer()
 
-        this.handleSubmitButtonClick = evt => {
-            evt.preventDefault()
+    this.handleSubmitButtonClick = evt => {
+      evt.preventDefault()
 
-            let {onSubmitButtonClick = helper.noop} = this.props
-            evt.resultString = this.resultString
-            onSubmitButtonClick(evt)
-        }
+      let {onSubmitButtonClick = noop} = this.props
+      evt.resultString = this.resultString
+      onSubmitButtonClick(evt)
     }
+  }
 
-    render({show, estimating, method, areaMap, board, komi, handicap}) {
-        if (isNaN(komi)) komi = 0
-        if (isNaN(handicap)) handicap = 0
+  render({show, estimating, method, areaMap, board, komi, handicap}) {
+    if (isNaN(komi)) komi = 0
+    if (isNaN(handicap)) handicap = 0
 
-        let score = areaMap && board && helper.getScore(board, areaMap, {handicap, komi})
-        let result = score && (method === 'area' ? score.areaScore : score.territoryScore)
+    let score = areaMap && board && getScore(board, areaMap, {handicap, komi})
+    let result =
+      score && (method === 'area' ? score.areaScore : score.territoryScore)
 
-        this.resultString = result > 0 ? t(p => `B+${p.result}`, {result})
-            : result < 0 ? t(p => `W+${p.result}`, {result: -result})
-            : t('Draw')
+    this.resultString =
+      result > 0 ? `B+${result}` : result < 0 ? `W+${-result}` : t('Draw')
 
-        return h(Drawer,
+    return h(
+      Drawer,
+      {
+        type: 'score',
+        show
+      },
+
+      h('h2', {}, t('Score')),
+
+      h(
+        'ul',
+        {class: 'tab-bar'},
+        h(
+          'li',
+          {class: classNames({current: method === 'area'})},
+          h(
+            'a',
             {
-                type: 'score',
-                show
+              href: '#',
+              onClick: this.handleAreaButtonClick
             },
-
-            h('h2', {}, 'Score'),
-
-            h('ul', {class: 'tabs'},
-                h('li', {class: classNames({current: method === 'area'})},
-                    h('a', {
-                        href: '#',
-                        onClick: this.handleAreaButtonClick
-                    }, t('Area'))
-                ),
-                h('li', {class: classNames({current: method === 'territory'})},
-                    h('a', {
-                        href: '#',
-                        onClick: this.handleTerritoryButtonClick
-                    }, t('Territory'))
-                )
-            ),
-
-            h('table', {},
-                h('thead', {}, h('tr', {},
-                    h('th'),
-                    h('th', {disabled: method === 'territory'}, t('Area')),
-                    h('th', {disabled: method === 'area'}, t('Territory')),
-                    h('th', {disabled: method === 'area'}, t('Captures')),
-                    h('th', {}, t('Komi')),
-                    h('th', {disabled: method === 'territory'}, t('Handicap')),
-                    h('th', {}, t('Total'))
-                )),
-                h('tbody', {},
-                    h(ScoreRow, {method, score, komi, handicap: 0, sign: 1}),
-                    h(ScoreRow, {method, score, komi, handicap, sign: -1})
-                )
-            ),
-
-            h('form', {},
-                h('p', {},
-                    t('Result:'), ' ',
-                    h('span', {class: 'result'}, this.resultString), ' ',
-
-                    !estimating && h('button', {
-                        type: 'submit',
-                        onClick: this.handleSubmitButtonClick
-                    }, t('Update Result')), ' ',
-
-                    h('button', {
-                        type: 'reset',
-                        onClick: this.handleCloseButtonClick
-                    }, t('Close'))
-                )
-            )
+            t('Area')
+          )
+        ),
+        h(
+          'li',
+          {class: classNames({current: method === 'territory'})},
+          h(
+            'a',
+            {
+              href: '#',
+              onClick: this.handleTerritoryButtonClick
+            },
+            t('Territory')
+          )
         )
-    }
-}
+      ),
 
-module.exports = ScoreDrawer
+      h(
+        'table',
+        {},
+        h(
+          'thead',
+          {},
+          h(
+            'tr',
+            {},
+            h('th'),
+            h('th', {disabled: method === 'territory'}, t('Area')),
+            h('th', {disabled: method === 'area'}, t('Territory')),
+            h('th', {disabled: method === 'area'}, t('Captures')),
+            h('th', {}, t('Komi')),
+            h('th', {disabled: method === 'territory'}, t('Handicap')),
+            h('th', {}, t('Total'))
+          )
+        ),
+        h(
+          'tbody',
+          {},
+          h(ScoreRow, {method, score, komi, handicap: 0, sign: 1}),
+          h(ScoreRow, {method, score, komi, handicap, sign: -1})
+        )
+      ),
+
+      h(
+        'form',
+        {},
+        h(
+          'p',
+          {},
+          t('Result:'),
+          ' ',
+          h('span', {class: 'result'}, this.resultString),
+          ' ',
+
+          !estimating &&
+            h(
+              'button',
+              {
+                type: 'submit',
+                onClick: this.handleSubmitButtonClick
+              },
+              t('Update Result')
+            ),
+          ' ',
+
+          h(
+            'button',
+            {
+              type: 'reset',
+              onClick: this.handleCloseButtonClick
+            },
+            t('Close')
+          )
+        )
+      )
+    )
+  }
+}
