@@ -1,23 +1,24 @@
-const GameTree = require('@sabaki/immutable-gametree')
-const sgf = require('@sabaki/sgf')
+import * as sgf from '@sabaki/sgf'
+import i18n from '../../i18n.js'
+import {getId} from '../helper.js'
+import * as gametree from '../gametree.js'
 
-const t = require('../../i18n').context('fileformats')
-const {getId} = require('../helper')
-const gametree = require('../gametree')
+const t = i18n.context('fileformats')
 
-exports.meta = {
-    name: t('Smart Game Format'),
-    extensions: ['sgf', 'rsgf']
+export const meta = {
+  name: t('Smart Game Format'),
+  extensions: ['sgf', 'rsgf']
 }
 
-let toGameTrees = rootNodes => rootNodes.map(root => gametree.new({getId, root}))
+let toGameTrees = rootNodes =>
+  rootNodes.map(root => gametree.new({getId, root}))
 
-exports.parse = function(content, onProgress = () => {}) {
-    let rootNodes = sgf.parse(content, {getId, onProgress})
-    return toGameTrees(rootNodes)
+export function parse(content, onProgress = () => {}) {
+  let rootNodes = sgf.parse(content, {getId, onProgress})
+  return toGameTrees(rootNodes)
 }
 
-exports.parseFile = function(filename, onProgress = () => {}) {
-    let rootNodes = sgf.parseFile(filename, {getId, onProgress})
-    return toGameTrees(rootNodes)
+export function parseFile(filename, onProgress = () => {}) {
+  let rootNodes = sgf.parseFile(filename, {getId, onProgress})
+  return toGameTrees(rootNodes)
 }
